Navigation and redirection
There was an issue where once you had clicked start tracking if you did not first select the cities tab before trying to do anything else it would cause bugs in the application. Now as soon as the index route is hit the navigate component will redirect the URL to app/cities preventing the user from not being on the right link.