Map Ground Work
I have added a functional map for this application using the leaflet and react-leaflet libraries.

I have begun laying the ground work for the maps interactivity. Right now selecting a city from the sidebar will move the map to that city as well as clicking somewhere which at the moment will bring up the form for adding a new location but as yet cannot add a new city.