Map functionality additions.
Today i have been working on the functionality of the map. To start i have reused a custom hook from an earlier project that i made for getting the geolocation from a users browser so that something can happen with it. I have used this for grabbing the data about where the user is clicking so that the form popup can be auto-filled and accurate. And allow the user to add cities and countries they have visited. Which is the functionality i will work on next.

I have linked the form to the incoming data and used a useEffect to get the necessary data from the API when the position changes and on mount.

I also added a button for this which allows the user to simply give permission for the application to place the marker where the user currently is, or atleast where their browser thinks they are.

I have added another custom hook for grabbing and using the lat, lng positions given in the URL i did this as there was now a need to use this code in multiple components and so i abstracted it into its own file for reusability.