Adding new cities.
Added: Ability to add new locations has been added. The user will click the form will come up with the location autofilled and once the user clicks create both the local state and the remote state are updated with the new cities array to keep them in sync. This change is reflected on screen after a brief load time.

Added: I also added the ability to delete cities from the list should the user realise they made a mistake or simply want to remove one for any reason they now can again both local and remote state are kept in sync with this change.

Refactor: I have refactored the CitiesContext to work with a reducer as i have come to learn this is a common combination in practice and overall i do think keeps things cleaner.