Login and Authentication
Over the last few days i have been working on implementing a "fake" authorization feature for logging in as a user. It is fake because the data is not stored in any 3rd party database but instead insecurely simply an object we are importing.

I created a new custom hook for the authorization and implemented it in the necessary areas. I also ran into bugs now with the user not being logged in but still trying to make use of the app and so i have redirected the user to the homepage when they are not signed in which prevents this issue and is how i feel it should work anyway.