Bug fixing and optimizations
I have fixed an issue where the getCity function could not be added to the dependency array for the useEffect in the context because this was creating an infinite loop, to fix this i have memoized the getCity function using the useCallback hook. which prevents unnecessary re-renders and fixes the issue.

I have optimized the application by implementing lazy loading the applications routes. Ensuring that the code is split and the pages are loaded when needed. I have also passed a fallback component to the suspense API to render a loading spinner while the page loads.